remove .slow from unit tests and extract formatter for values
